requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
boto3==1.34.40
botocore==1.34.40
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.2.0
Django==5.0.7
django-filter==23.5
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
drf-yasg==1.21.7
Faker==26.0.0
flake8==7.0.0
gunicorn==21.2.0
inflection==0.5.1
isort==5.13.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
mccabe==0.7.0
minio==7.2.3
packaging==23.2
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.19.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
